Make count in get_audio usize

+++ src/lib.rs
@@ -734,7 +734,7 @@ impl MmC2r<'_> {
         Ok(mm)
     }
     /// Fill a buffer with delicious samples
-    pub fn get_audio(&mut self, output_buffer: &mut [i16], mut count: i64) -> bool {
+    pub fn get_audio(&mut self, output_buffer: &mut [i16], mut count: usize) -> bool {
         if self.src.num_channels <= 0 {
             return false;
         }
@@ -742,8 +742,8 @@ impl MmC2r<'_> {
         let mut cnt = true;
         while count > 0 {
             let mut remain = self.playback.tick_len - self.playback.tick_offset;
-            if remain > count {
-                remain = count;
+            if remain > count as i64 {
+                remain = count as i64;
             }
             let mut chan_idx = 0;
             while chan_idx < self.src.num_channels {
@@ -764,7 +764,7 @@ impl MmC2r<'_> {
                 self.playback.tick_offset = 0;
             }
             offset += remain;
-            count -= remain;
+            count -= remain as usize;
         }
         cnt
     }
